Fix filter method after rolltable generation
Don't know if this changed over time. 
But the original check didn't work.

The weight property should be checked for `0` instead of checking if the entity is undefined.
This error only occurred when using a `weightPredicate` and this predicate would return 0 (as per example from the Wiki) for an entity.